Fix overflow: use ll instead of int in fenwick_tree

+++ content/data-structures/bit.cpp
@@ -4,22 +4,22 @@
  *Status:* Tested
 */
 struct fenwick_tree {
-  vector <int> bit; int n;
+  vector<ll> bit; int n;
   fenwick_tree(int n): n(n) { bit.assign(n, 0); }
-  fenwick_tree(vector <int> &a): fenwick_tree(a.size()) {
+  fenwick_tree(vector<ll> &a): fenwick_tree(a.size()) {
     for (size_t i = 0; i < a.size(); i++)
       add(i, a[i]);
   }
-  int sum(int r) {
-    int ret = 0;
+  ll sum(int r) {
+    ll ret = 0;
     for (; r >= 0; r = (r & (r + 1)) - 1)
       ret += bit[r];
     return ret;
   }
-  int sum(int l, int r) {
+  ll sum(int l, int r) {
     return sum(r) - sum(l - 1);
   }
-  void add(int idx, int delta) {
+  void add(int idx, ll delta) {
     for (; idx < n; idx = idx | (idx + 1))
       bit[idx] += delta;
   }
